Refactor misc
